WIP: Implement preview screenshots (crashes atm)

+++ src/openrct2/park/ParkPreview.h
@@ -29,8 +29,8 @@ namespace OpenRCT2
     struct PreviewImage
     {
         PreviewImageType type;
-        uint8_t width;
-        uint8_t height;
+        uint16_t width;
+        uint16_t height;
         uint8_t pixels[250 * 250]{};
     };
